+++ docs/wksp/05-semantic-kernel-workshop/simple-ai-orchestration/index.md
@@ -7,8 +7,8 @@ chatbot as a console application:
 1. [Lesson 1: Semantic Kernel chatbot](lesson1.md)
 1. [Lesson 2: Semantic Kernel chatbot with history](lesson2.md)
 1. [Lesson 3: Semantic Kernel chatbot with plugins](lesson3.md)
-1. [Lesson 4: Semantic Kernel chatbot with Chat Completion Agent](lesson4.md)
-1. [Lesson 5: Describe all plugins in Semantic Kernel chatbot](lesson5.md)
+1. [Lesson 4: Describe all plugins in Semantic Kernel chatbot](lesson4.md)
+1. [Lesson 5: Semantic Kernel chatbot with Chat Completion Agent](lesson5.md)
 1. [Lesson 6: Create Semantic Kernel chat completion agent](lesson6.md)
 
 Subsequently the following lessons guide you through converting that chatbot application
