+++ workshop/dotnet/Lessons/Lesson4/Program.cs
@@ -3,10 +3,10 @@ using Core.Utilities.Config;
 using Core.Utilities.Plugins;
 // Add import required for StockService
 using Core.Utilities.Services;
+// TODO: Step 1 - Add import for ModelExtensionMethods
+
 using Microsoft.SemanticKernel;
 using Microsoft.SemanticKernel.Connectors.OpenAI;
-// TODO: Step 1 - Add import for Agents
-
 // Add ChatCompletion import
 using Microsoft.SemanticKernel.ChatCompletion;
 // Temporarily added to enable Semantic Kernel tracing
@@ -28,8 +28,6 @@ HttpClient httpClient = new();
 StockDataPlugin stockDataPlugin = new(new StocksService(httpClient));
 kernel.Plugins.AddFromObject(stockDataPlugin);
 
-// TODO: Step 2 - Add code to create Stock Sentiment Agent
-
 // Get chatCompletionService and initialize chatHistory with system prompt
 var chatCompletionService = kernel.GetRequiredService<IChatCompletionService>();
 ChatHistory chatHistory = new("You are a friendly financial advisor that only emits financial advice in a creative and funny tone");
@@ -44,6 +42,9 @@ OpenAIPromptExecutionSettings promptExecutionSettings = new()
 // Initialize kernel arguments
 KernelArguments kernelArgs = new(promptExecutionSettings);
 
+// TODO: Step 2 - add call to print all plugins and functions
+
+// TODO: Step 3 - Comment out all code after "Execute program" comment
 // Execute program.
 const string terminationPhrase = "quit";
 string? userInput;
@@ -59,7 +60,7 @@ do
         string fullMessage = "";
         chatHistory.AddUserMessage(userInput);
 
-        // TODO: Step 3 - Replace chatCompletionService with stockSentimentAgent
+        // Provide promptExecutionSettings and kernel arguments
         await foreach (var chatUpdate in chatCompletionService.GetStreamingChatMessageContentsAsync(chatHistory, promptExecutionSettings, kernel))
         {
             Console.Write(chatUpdate.Content);
